KG-Enhanced-QA/requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.1.17
langchain-openai==0.0.7
openai==1.3.5
py2neo==2021.2.4
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
graphviz==0.20.1
tiktoken==0.6.0
SQLAlchemy==2.0.23